print_sigevent: remove redundant cast
* print_sigevent.c (print_sigevent): Remove redundant cast from
struct_sigevent.sigev_value.sival_ptr to unsigned long as the former
is not a pointer but an unsigned integer.

+++ print_sigevent.c
@@ -46,7 +46,7 @@ MPERS_PRINTER_DECL(void, print_sigevent, struct tcb *tcp, const long addr)
 	if (sev.sigev_value.sival_ptr) {
 		tprintf("sigev_value={sival_int=%d, sival_ptr=",
 			sev.sigev_value.sival_int);
-		printaddr((unsigned long) sev.sigev_value.sival_ptr);
+		printaddr(sev.sigev_value.sival_ptr);
 		tprints("}, ");
 	}
